+++ CHANGELOG.md
@@ -48,6 +48,7 @@ The sliding behavior in `o-slider` component requires that HammerJS is loaded on
 
 ### Bug Fixes
 * **o-form-navigation**: fixing index error ([f487e54](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/f487e54)), closes [#238](https://github.com/OntimizeWeb/ontimize-web-ngx/issues/238)
+* **OServiceBaseComponent**: fixed bug on service components with pagination and located inside a detail form ([9c69c4b](https://github.com/OntimizeWeb/ontimize-web-ngx/commit/9c69c4b))
 
 ### BREAKING CHANGES
 * **o-row**, **o-column**: